Changed the for loops in the parse functions to for… of… loops.
The parse functions for those with APIs assume that the length of the result list returned by the API will always be equal to the number of results requested, which isn’t always the case. This is evident in the for loop which iterates until the number of results requested instead of the length of the list. So, all those for loops have been changed to for… of… loops which is far more readable and requires less lines of code, but it’ll be very slightly slower.