Moved everything to the src folder.
Moved all the files to be built into a source folder and excluded the alternate-router folder in the tsconfig.json so that the application will build successfully. As such, the migrate directory has been removed and replaced with the alternate-router folder.

Added a shell script to easily change the routers so the migration process is easier.

Added a function to change the searchParams back into a query string in the app/search/page.tsx file so that the parseQuery function can handle it.