Removed the dependency on the router in the components.
Removed the dependency on the router within all the search page components, like the search bar, the page selector and the search page. Now the router is passed to the component through the page so that it works with both the app router and the pages router.

Added an interface called Router in the types file.