Fixed the license part of the footer
Changed the license part of the footer to have the copyleft notice above the "All Wrongs Reserved", which is much easier to deal with than having both in a row.

Changed the margin on the icon on the main page to a gap, and made that gap zero when on mobile.

Also changed the link for the license file from GitHub to Codeberg.