Removed the wait function
The wait function is no longer needed now the transition for visibility is the same as that of the opacity, which makes the button fade in and out without the need for the JavaScript wait function which works poorly compared to the CSS transition property.

Somehow, the first result when searching for "autism" causes the first page of the search results to overflow. This causes the webpage to just go absolutely bonkers on mobile as it is trying to display the page with the width of the desktop on mobile. It is apparently caused by the bolded text to highlight the search term, which somehow causes the text in the list to overflow, despite the text being able to wrap. This needs to be looked into more.