Fixed the main page in landscape mode
Added a fixed with to the research icon on the main page as Safari doesn’t know how to calculate the width of the research icon properly when only given the height for some weird reason.

Made the div that marks the top of the page completely invisible since there’s no content inside it.